+++ docs/CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.29] - 2026-08-03
+
+### Added
+
+- Option `--manifest-file` for `queue download` and `playlist download`, setting the
+  path of the download manifest with the `{output_directory}` and `{timestamp}`
+  placeholders
+
+### Changed
+
+- The download manifest of `queue download` and `playlist download` is written to
+  `manifest.json` (was `queue.json`) inside the output directory by default
+
+
 ## [0.0.28] - 2026-08-03
 
 ### Changed